+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGE LOG #
 
+July 20th 2013
+* removed netbeans files from repo ;)
+* comment fixes
+* 4-full-mvc-framework: new user verification mail does no work with user_id, not user_email
+* 2-advanced: new user verification mail does no work with user_id, not user_email
+(to prevent weird URL encoding problems with special email adresses)
+
+
 July 16th 2013
 * 4-full-mvc-framework: captcha is now case-insensitive
 * 2-advanced: captcha is now case-insensitive